240426 article modify

+++ user-front/app/component/articles/modul/columns.tsx
@@ -31,9 +31,7 @@ export default function articleColumns(): GridColDef[] {
             field: 'title',
             headerName: 'TITLE',
             renderCell: ({ row }: CellType) =>
-                MyTypography(
-                    <Link href={`${PG.ARTICLE}/detail/${row.id}`}> {row.title} </Link>
-                    , "1.5rem")
+                MyTypography(<Link href={`${PG.ARTICLE}/detail/${row.id}`}> {row.title} </Link> , "1.5rem")
         },
         {
             flex: 0.04,
@@ -71,27 +69,20 @@ export default function articleColumns(): GridColDef[] {
             flex: 0.04,
             minWidth: 30,
             sortable: false,
-            field: 'update',
-            headerName: 'UPDATE',
-            renderCell: ({ row }: CellType) =>
-                <button className="btn overflow-hidden relative w-full h-full bg-blue-500 text-white rounded-xl font-bold uppercase -- before:block before:absolute before:h-full before:w-1/2 before:rounded-full 
-            before:bg-pink-400 before:top-0 before:left-1/4 before:transition-transform before:opacity-0 before:hover:opacity-100 hover:text-200 hover:before:animate-ping transition-all duration-300"
-                    onClick={() => `${PG.ARTICLE}/update/${row.id}`}> UPDATE</button>   
-        },
-        {
-            flex: 0.04,
-            minWidth: 30,
-            sortable: false,
             field: 'delete',
             headerName: 'DELETE',
             renderCell: ({ row }: CellType) =>
-                <button className="btn overflow-hidden relative w-full h-full bg-blue-500 text-white rounded-xl font-bold uppercase -- before:block before:absolute before:h-full before:w-1/2 before:rounded-full 
+                <button className="btn focus overflow-hidden relative w-full h-full bg-blue-500 text-white rounded-xl font-bold uppercase -- before:block before:absolute before:h-full before:w-1/2 before:rounded-full 
             before:bg-pink-400 before:top-0 before:left-1/4 before:transition-transform before:opacity-0 before:hover:opacity-100 hover:text-200 hover:before:animate-ping transition-all duration-300"
                     onClick={() => {
-                        alert("article을 삭제합니다.")
-                        console.log("delete article id : {}", row.id)
-                        dispatch(deleteArticleById(row.id))
-                        location.reload();
+                        let flag = confirm("article을 삭제하시겠습니까?")
+                        if(flag){
+                            console.log("delete article id : {}", row.id)
+                            dispatch(deleteArticleById(row.id))
+                            location.reload();
+                        }else{  
+                            alert("article 삭제가 취소되었습니다.")
+                        }
                     }
                     }> DELETE</button>
         }
